refactor(node): split txindex query modules
